requirements.txt
scipy
numpy
matplotlib
Pillow
gdown
tensorflow-gpu==2.4.0
pandas
